Enable fit powerlaw to download from wandb directly

+++ src/genomenlp/fit_powerlaw.py
@@ -15,6 +15,7 @@ import matplotlib.pyplot as plt
 from transformers import \
     AutoModel, BertModel, DistilBertModel, RobertaModel, XLNetModel
 from utils import plot_hist, plot_scatter, _cite_me
+import wandb
 import weightwatcher as ww
 
 def main():
@@ -49,22 +50,37 @@ def main():
     # logger = logging.getLogger('weightwatcher')
     # logger.setLevel(logging.INFO)
     # warnings.filterwarnings('ignore')
+
+    model_path_final = list()
+
     if model_path != None:
         for i in model_path:
-            model = AutoModel.from_pretrained(i)
-            watcher = ww.WeightWatcher(model=model)
-            details = watcher.describe()
-            print("\nMODEL_DETAILS (summary):\n")
-            print(details)
-            model_out = "/".join([output_dir, os.path.basename(i)])
-            details = watcher.analyze(
-                randomize=True, min_evals=50, plot=True, savefig=model_out
-                )
-            print("\nMODEL_DETAILS (with fit):\n")
-            print(details)
-            alpha_main = "/".join([model_out, "alpha_main.tsv"])
-            details.to_csv(alpha_main, sep="\t")
-            model_info = details[(details.alpha<alpha_max) & (details.alpha>0)]
+            if os.path.exists(i):
+                print("Use model files from:", i)
+                model_path_final.append(i)
+            else:
+                print("Downloading model files from wandb:", i)
+                api = wandb.Api(timeout=10000)
+                run = api.run(path=i)
+                for j in run.files():
+                    i = "/".join([output_dir, os.path.basename(i)])
+                    j.download(root=i, replace=True) 
+                print("Loading model from:", i)
+                model = AutoModel.from_pretrained(i)
+                watcher = ww.WeightWatcher(model=model)
+                details = watcher.describe()
+                print("\nMODEL_DETAILS (summary):\n")
+                print(details)
+                model_out = "/".join([output_dir, os.path.basename(i)])
+                model_path_final.append(model_out)
+                details = watcher.analyze(
+                    randomize=True, min_evals=50, plot=True, savefig=model_out
+                    )
+                print("\nMODEL_DETAILS (with fit):\n")
+                print(details)
+                alpha_main = "/".join([model_out, "alpha_main.tsv"])
+                details.to_csv(alpha_main, sep="\t")
+                model_info = details[(details.alpha<alpha_max) & (details.alpha>0)]   
 
     alpha_hist = "/".join([output_dir, "alpha_hist.pdf"])
     alpha_plot = "/".join([output_dir, "alpha_plot.pdf"])
@@ -73,7 +89,7 @@ def main():
     if model_path != None:
         print("\nCOMPARISONS:\n", model_path, "\n")
         model_path = [(i,ww.WeightWatcher(AutoModel.from_pretrained(i)).analyze(
-                        randomize=True, min_evals=50)) for i in model_path]
+                        randomize=True, min_evals=50)) for i in model_path_final]
         for i, j in model_path:
             out_dir = "".join([output_dir, "/", i.split("/")[-1], "/"])
             if not os.path.isdir(out_dir):
